working on selection highlighting
I'm getting a GL_INVALID_VALUE error, and I haven't found the source
yet.  Edited the shaders to use a texture for the color, and included
the SOIL library for loading in the textures.  To run it in VS, make
sure you include SOIL.lib in additional dipendency.  I'll try and find
the GL error when I get home from work.